Copy data to the volume from the entrypoint
Previously, data was copied to the volume inside the Dockerfile using
a COPY statement, but support for this is implementation-dependent at
best (and it breaks alternative building tools like buildah).

Another noteworthy point is that configuration files are now
written conditionally on a per-file basis (instead of on the simple
/etc/opensemanticsearch/version sentinel). Fortunately cp -n takes care
of the heavy lifting for us.

Migration and entity generation is also run unconditionally now, so that
changes to the database are picked up automatically.